Extract useOnClickOutside
