+++ helpers/useOnClickOutside.ts
@@ -1,0 +1,15 @@
+import React, { useEffect } from "react"
+
+export function useOnClickOutside(ref: React.RefObject<HTMLDivElement>, cb: Function) {
+  useEffect(() => {
+    function handleClickOutside(event: MouseEvent) {
+      if (ref.current && !ref.current.contains(event.target as Node)) {
+        cb()
+      }
+    }
+    document.addEventListener('mousedown', handleClickOutside)
+    return () => {
+      document.removeEventListener('mousedown', handleClickOutside)
+    }
+  }, [])
+}
